Fix commands allowing any string with command prefix text.
Commands were allowing any string to count as the command as
long as the starting characters were equal to the command text.
This fix instead uses strlen(argv[0]) to check against the user
input text instead of the command string, ensuring a command
only runs if the input is for that command, not any text starting
with the command.

Credit and thanks to YT viewer _AB_ for this fix.