editor.c: Fix page fault when loading a file, fix saving txt file
Page fault from returning 0 bytes read from read() syscall is
bypassed by rewinding the file before the initial read() when
loading the file. This is not a fix per se, as it should not page
fault or crash when read() returns 0 bytes, still need to fix that.

Saving with CTRL-S rewinds the file first to overwrite all contents.